Убирает формирование данных из модели, немного рефакторит, исправляет ошибку в фугкции расчета остаточного времени
