+++ src/const.js
@@ -1,8 +1,3 @@
 const POINTS_TYPE = ['Taxi', 'Ship', 'Flight'];
 
-const TimeForDate = {
-  MIN_IN_HOUR: 60,
-  MIN_IN_MONTH: 1440,
-};
-
-export { POINTS_TYPE, TimeForDate };
+export { POINTS_TYPE };
