+++ src/mock/const.js
@@ -1,6 +1,12 @@
 const DESCRIPTION = 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Cras aliquet varius magna, non porta ligula feugiat eget. Fusce tristique felis at fermentum pharetra. Aliquam id orci ut lectus varius viverra. Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. Sed sed nisi sed augue convallis suscipit in sed felis. Aliquam erat volutpat. Nunc fermentum tortor ac porta dapibus. In rutrum ac purus sit amet tempus.';
 const PHOTO_URLS = 'https://loremflickr.com/248/152?random=';
-const POINTS_COUNT = 4;
-const CITIES = ['Istambul', 'Moscow', 'Califormia', 'Kazan', 'Vladivostok', 'Usuriysk', 'Tihvan'];
+const POINTS_COUNT = 10;
+const CITIES = ['Istanbul', 'Moscow', 'California', 'Kazan', 'Vladivostok', 'Usuriysk', 'Tihvan'];
+const Time = {
+  START:['2023-08-06T00:05:00.845Z', '2023-08-07T01:10:20.845Z', '2023-08-08T03:19:43.845Z', '2023-08-09T04:16:32.845Z', '2023-08-06T05:28:12.845Z',],
+  END: ['2023-08-11T10:30:20.845Z', '2023-08-12T11:13:21.845Z', '2023-08-13T12:00:02.845Z', '2023-08-14T13:42:32.845Z', '2023-08-15T14:53:52.845Z',],
+};
 
-export { DESCRIPTION, PHOTO_URLS, POINTS_COUNT, CITIES };
+const PRICE = [10000, 100, 20, 15, 12345, 520, 9999, 321, 500, 1212, 1010, 2015, 3050];
+
+export { DESCRIPTION, PHOTO_URLS, POINTS_COUNT, CITIES, Time, PRICE };
