+++ src/view/point-view.js
@@ -1,170 +1,82 @@
 import AbstractView from '../framework/view/abstract-view.js';
-import { filterDateForNewPoint } from '../utils';
-import { getRandomArrayElement } from '../mock/utils';
-import { POINTS_TYPE } from '../const';
+import {
+  filterHoursPoints,
+  getTimeInterval,
+  filterDayMonth,
+  filterPointDay,
+} from '../utils.js';
 import { mockOffers } from '../mock/points.js';
 
-const BLANK_DATA_TRIP = {
-  basePrice: 0,
-  typePoint: getRandomArrayElement(POINTS_TYPE),
-  destination: null,
-  timeStart: null,
-  timeEnd: null,
-  offersCheck: [],
-  description: '',
-  photos: null,
-  isFavorite: false,
-};
-
 function createBoardTemplate(data) {
-  const { typePoint, destination, timeStart, timeEnd, offersCheck, description, photos } = data;
+  const { typePoint, destination, timeStart, timeEnd, offersCheck, basePrice } = data;
 
   const typeOffersObj = mockOffers.find((item) => item.type === typePoint);
   const { offers } = typeOffersObj;
 
-  const createTypeOffersTemplate = offers.map(({ id, title, price }) => `
-  <div class="event__offer-selector">
-    <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox" name="event-offer-luggage" ${offersCheck.map((elem) => {
-    if (elem === id) {
-      return 'checked';
+  const normalizeTimeStart = filterHoursPoints(timeStart);
+  const normalizeTimeEnd = filterHoursPoints(timeEnd);
+  const normalizeDate = filterDayMonth(timeStart);
+
+  return /*html*/ `
+  <li class="trip-events__item">
+    <div class="event">
+      <time class="event__date" datetime="${filterPointDay(
+    timeStart
+  )}">${normalizeDate}</time>
+      <div class="event__type">
+      <img class="event__type-icon" width="42" height="42" src="img/icons/${typePoint}.png" alt="Event type icon">
+    </div>
+    <h3 class="event__title">${typePoint} ${destination}</h3>
+      <div class="event__schedule">
+        <p class="event__time">
+          <time class="event__start-time" datetime="2019-03-19T18:00">${normalizeTimeStart}</time>
+          &mdash;
+          <time class="event__end-time" datetime="2019-03-19T19:00">${normalizeTimeEnd}</time>
+        </p>
+      <p class="event__duration">${getTimeInterval(timeStart, timeEnd)}</p>
+    </div>
+    <p class="event__price">
+      &euro;&nbsp;<span class="event__price-value">${basePrice}</span>
+    </p>
+    <h4 class="visually-hidden">Offers:</h4>
+    <ul class="event__selected-offers">
+      ${offers.map(({ id, title, price }) => offersCheck.map((checkedOfferId) => {
+    if (checkedOfferId === id) {
+      return `<li class="event__offer">
+              <span class="event__offer-title">${title}</span>
+              &plus;&euro;&nbsp;
+              <span class="event__offer-price">${price}</span>
+            </li>`;
     }
-  }).join('')} >
-    <label class="event__offer-label" for="event-offer-luggage-1">
-    <span class="event__offer-title">${title}</span>
-    &plus;&euro;&nbsp;
-    <span class="event__offer-price">${price}</span>
-    </label>
-  </div>`).join('');
-
-  return /*html*/ `<li class="trip-events__item">
-  <form class="event event--edit" action="#" method="post">
-    <header class="event__header">
-      <div class="event__type-wrapper">
-        <label class="event__type  event__type-btn" for="event-type-toggle-1">
-          <span class="visually-hidden">Choose event type</span>
-          <img class="event__type-icon" width="17" height="17" src="img/icons/${typePoint}.png" alt="Event type icon">
-        </label>
-        <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
-
-        <div class="event__type-list">
-          <fieldset class="event__type-group">
-            <legend class="visually-hidden">Event type</legend>
-
-            <div class="event__type-item">
-              <input id="event-type-taxi-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="taxi">
-              <label class="event__type-label  event__type-label--taxi" for="event-type-taxi-1">Taxi</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-bus-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="bus">
-              <label class="event__type-label  event__type-label--bus" for="event-type-bus-1">Bus</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-train-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="train">
-              <label class="event__type-label  event__type-label--train" for="event-type-train-1">Train</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-ship-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="ship">
-              <label class="event__type-label  event__type-label--ship" for="event-type-ship-1">Ship</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-drive-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="drive">
-              <label class="event__type-label  event__type-label--drive" for="event-type-drive-1">Drive</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-flight-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="flight" checked>
-              <label class="event__type-label  event__type-label--flight" for="event-type-flight-1">Flight</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="check-in">
-              <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-sightseeing-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="sightseeing">
-              <label class="event__type-label  event__type-label--sightseeing" for="event-type-sightseeing-1">Sightseeing</label>
-            </div>
-
-            <div class="event__type-item">
-              <input id="event-type-restaurant-1" class="event__type-input  visually-hidden" type="radio" name="event-type" value="restaurant">
-              <label class="event__type-label  event__type-label--restaurant" for="event-type-restaurant-1">Restaurant</label>
-            </div>
-          </fieldset>
-        </div>
-      </div>
-
-      <div class="event__field-group  event__field-group--destination">
-        <label class="event__label  event__type-output" for="event-destination-1">
-          ${typePoint}
-        </label>
-        <input class="event__input  event__input--destination" id="event-destination-1" type="text" name="event-destination" value="${destination}" list="destination-list-1">
-        <datalist id="destination-list-1">
-          <option value="Amsterdam"></option>
-          <option value="Geneva"></option>
-          <option value="Chamonix"></option>
-        </datalist>
-      </div>
-
-      <div class="event__field-group  event__field-group--time">
-        <label class="visually-hidden" for="event-start-time-1">From</label>
-        <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time" value="${filterDateForNewPoint(timeStart)}">
-        &mdash;
-        <label class="visually-hidden" for="event-end-time-1">To</label>
-        <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time" value="${filterDateForNewPoint(timeEnd)}">
-      </div>
-
-      <div class="event__field-group  event__field-group--price">
-        <label class="event__label" for="event-price-1">
-          <span class="visually-hidden">Price</span>
-          &euro;
-        </label>
-        <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="">
-      </div>
-
-      <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
-      <button class="event__reset-btn" type="reset">Cancel</button>
-    </header>
-    <section class="event__details">
-      <section class="event__section  event__section--offers">
-        <h3 class="event__section-title  event__section-title--offers">Offers</h3>
-
-        <div class="event__available-offers">
-
-        ${createTypeOffersTemplate}
-
-      </section>
-
-      <section class="event__section  event__section--destination">
-        <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-        <p class="event__destination-description">${description}</p>
-
-        <div class="event__photos-container">
-          <div class="event__photos-tape">
-          ${photos.map((el) => `<img class="event__photo" src="${el}" alt="Event photo">`)}
-          </div>
-        </div>
-
-      </section>
-    </section>
-  </form>
-</li>`;
+  }).join('')).join('')}
+    </ul >
+    <button class="event__favorite-btn" type="button">
+      <span class="visually-hidden">Add to favorite</span>
+      <svg class="event__favorite-icon" width="28" height="28" viewBox="0 0 28 28">
+        <path d="M14 21l-8.22899 4.3262 1.57159-9.1631L.685209 9.67376 9.8855 8.33688 14 0l4.1145 8.33688 9.2003 1.33688-6.6574 6.48934 1.5716 9.1631L14 21z"/>
+      </svg>
+    </button>
+    <button class="event__rollup-btn" type="button">
+      <span class="visually-hidden">Open event</span>
+    </button>
+  </div >
+</li > `;
 }
 
-
-export default class NewPoint extends AbstractView {
-  #data = null;
-
-  constructor(data = BLANK_DATA_TRIP) {
+export default class TripEventsItem extends AbstractView {
+  constructor(data) {
     super();
     this.data = data;
+    // this.handleClick = onClick;
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#clickHandler);
   }
 
   get template() {
     return createBoardTemplate(this.data);
   }
+
+  #clickHandler = (evt) => {
+    evt.preventDefault();
+    // this.handleClick();
+  };
 }
